Bring all the changes from fresh_quill_extensions
