+++ flutter_quill_extensions/lib/core/exceptions.dart
@@ -1,0 +1,12 @@
+// import 'package:meta/meta.dart';
+
+// @immutable
+// class NetworkException implements Exception {
+//   const NetworkException({required this.message});
+
+//   final String message;
+
+//   @override
+//   String toString() =>
+//       'Error while loading something from the network: $message';
+// }
