+++ flutter_quill_extensions/lib/logic/services/image_saver.dart
@@ -1,0 +1,7 @@
+abstract class ImageSaverInterface {
+  const ImageSaverInterface();
+  Future<void> saveLocalImage(String imageUrl);
+  Future<void> saveImageFromNetwork(Uri imageUrl);
+  Future<bool> hasAccess({required bool toAlbum});
+  Future<bool> requestAccess({required bool toAlbum});
+}
