+++ flutter_quill_extensions/lib/presentation/embeds/widgets/simple_dialog_item.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class SimpleDialogItem extends StatelessWidget {
+  const SimpleDialogItem({
+    required this.icon,
+    required this.color,
+    required this.text,
+    required this.onPressed,
+    super.key,
+  });
+
+  final IconData icon;
+  final Color color;
+  final String text;
+  final VoidCallback onPressed;
+
+  @override
+  Widget build(BuildContext context) {
+    return SimpleDialogOption(
+      onPressed: onPressed,
+      child: Row(
+        children: [
+          Icon(icon, size: 36, color: color),
+          Padding(
+            padding: const EdgeInsetsDirectional.only(start: 16),
+            child:
+                Text(text, style: const TextStyle(fontWeight: FontWeight.bold)),
+          ),
+        ],
+      ),
+    );
+  }
+}
