+++ flutter_quill_extensions/lib/presentation/models/config/editor/webview.dart
@@ -1,0 +1,6 @@
+import 'package:meta/meta.dart' show immutable;
+
+@immutable
+class QuillEditorWebViewEmbedConfigurations {
+  const QuillEditorWebViewEmbedConfigurations();
+}
