+++ flutter_quill_extensions/pubspec.yaml
@@ -1,8 +1,13 @@
 name: flutter_quill_extensions
 description: Embed extensions for flutter_quill including image, video, formula and etc.
-version: 0.5.1
+version: 1.0.0-dev
 homepage: https://bulletjournal.us/home/index.html
 repository: https://github.com/singerdmx/flutter-quill/tree/master/flutter_quill_extensions
+topics:
+ - ui
+ - widgets
+ - widget
+ - rich-text-editor
 platforms:
   android:
   ios:
@@ -12,31 +17,35 @@ platforms:
   windows:
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
-  flutter: ">=3.0.0"
+  sdk: '>=3.1.3 <4.0.0'
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_quill: ^7.8.0
+  flutter_quill: ^8.1.1
   # In case you are working on changes for both libraries,
   # flutter_quill:
-  #   path: ../
+  #   path: ./../flutter-quill
 
   http: ^1.1.0
-  image_picker: ">=1.0.4"
   photo_view: ^0.14.0
-  video_player: ^2.7.2
+  video_player: ^2.8.1
   youtube_player_flutter: ^8.1.2
-  math_keyboard: ">=0.2.1"
+  flutter_inappwebview: ^5.8.0
   universal_html: ^2.2.4
-
   gal: ^2.1.2
+  path: ^1.8.3
+  image_picker: ^1.0.4
+  math_keyboard: ^0.2.1
+  url_launcher: ^6.2.1
+  meta: ^1.9.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.11.1
-
+  flutter_lints: ^3.0.0
+  
 flutter:
+  uses-material-design: true
